+++ src/auth/HomePrivate.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Outlet, Navigate } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { RootState } from 'states/store';
-import AdminHomePage from 'components/admin-home/AdminHomePage';
-import NoGroupPage from 'components/admin-home/NoGroupPage';
+import AdminHomePage from 'components/admin-MainPage/AdminHomePage';
+import NoGroupPage from 'components/admin-MainPage/NoGroupPage';
 import OnBoardingPage from 'pages/OnBoardingPage';
 
 const HomePrivate = () => {
